initial implementation working on test bench
