+++ platformio.ini
@@ -15,7 +15,10 @@ default_envs = teensy41
 platform = teensy
 board = teensy41
 framework = arduino
-lib_deps = nanopb/Nanopb@^0.4.7
+lib_deps = 
+    nanopb/Nanopb@^0.4.7
+    adafruit/Adafruit GFX Library@^1.11.9
+	adafruit/Adafruit SSD1306@^2.5.7
 test_filter = hardware/*
 custom_nanopb_protos = +<lib/PIER/log_message.proto>, +<lib/PIER/header_message.proto>
 custom_nanopb_python_loc = data
